fix `cargo test` (broken by reorg)
due to changes in the structure, one test had to be moved from
`rep_lang_parser` to `rep_lang_runtime`. then, due to restrictions on no orphan
instances (I forget what Rust calls them, but that's what Haskell calls them),
I had to move some things around.
