+++ rep_lang_parser/Cargo.toml
@@ -9,3 +9,8 @@ edition = "2018"
 [dependencies]
 combine = "4.3.2"
 pretty = "0.10.0"
+quickcheck = "0.9.2"
+rand = "0.7"
+
+[dev-dependencies]
+quickcheck_macros = "0.9.1"
