+++ rep_lang_parser/src/test.rs
@@ -1,5 +1,2 @@
 #[cfg(test)]
 pub mod parse_pretty;
-
-#[cfg(test)]
-pub mod syntax;
