+++ rep_lang_runtime/src/lib.rs
@@ -1,6 +1,15 @@
+// TODO should this go here?
+#[cfg(test)]
+extern crate quickcheck;
+#[cfg(test)]
+#[macro_use(quickcheck)]
+extern crate quickcheck_macros;
+
 pub mod env;
 pub mod eval;
 pub mod infer;
 pub mod toplevel;
 pub mod types;
 pub mod types_values;
+
+pub mod test;
