+++ rep_lang_runtime/src/test.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod syntax;
