+++ rep_lang_runtime/src/test/syntax.rs
@@ -1,0 +1,17 @@
+use quickcheck::quickcheck;
+
+use rep_lang_parser::{
+    syntax::{primop_arity, PrimOp},
+};
+
+use crate::{
+    infer::{infer_primop, InferState},
+    types::type_arity,
+};
+
+#[quickcheck]
+fn primop_arity_eql(op: PrimOp) -> bool {
+    let mut is = InferState::new();
+    let ty = infer_primop(&mut is, &op);
+    primop_arity(&op) == type_arity(ty)
+}
